Showing another book, if first book is on loan
There was a problem where the app would show a book that was on loan
even when there were one or more other books available.

Now the app prefers books that are available over those that are on
loan.

If all books are on loan the first will be shown which will be the
newest.
The preference for the newest books as before is unchanged.